meetwise-runpod/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
pyannote.audio==3.1.1
torch==2.1.0
torchaudio==2.1.0
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.4
transformers==4.35.2
datasets==2.14.6
accelerate==0.24.1 